[simulation.view] Work in changes requested. Decouple tank width and height.

+++ src/osip-simulation-view/src/main/java/edu/kit/pse/osip/simulation/view/main/Drawer.java
@@ -5,6 +5,9 @@ import javafx.scene.canvas.GraphicsContext;
 /**
  * The interface from which all other GUI elements originate. It specifies the draw() method with
  * which the element draws itself onto the screen according to its implementation.
+ *
+ * @version 1.0
+ * @author Niko Wilhelm
  */
 public interface Drawer {
     /**
@@ -12,5 +15,5 @@ public interface Drawer {
      * @param context The context that the object draws itself onto
      * @param time The current time in nanoseconds as supplied by ApplicationTimer
      */
-    void draw(GraphicsContext context, long time);
+    void draw(GraphicsContext context, double time);
 }
