+++ src/osip-simulation-view/src/main/java/edu/kit/pse/osip/simulation/view/main/TemperatureSensorDrawer.java
@@ -8,6 +8,9 @@ import javafx.scene.paint.Color;
 
 /**
  * This class visualizes a temperature sensor
+ *
+ * @version 1.0
+ * @author Niko Wilhelm
  */
 public class TemperatureSensorDrawer extends ObjectDrawer {
 
@@ -32,9 +35,10 @@ public class TemperatureSensorDrawer extends ObjectDrawer {
     /**
      * The Drawer draws itself onto the GraphicsContext at its getPosition().
      * @param context The context that the object draws itself onto
+     * @param currentTime
      */
     @Override
-    public final void draw(GraphicsContext context, long currentTime) {
+    public final void draw(GraphicsContext context, double currentTime) {
         Canvas canvas = context.getCanvas();
 
         double temp = (tank.getLiquid().getTemperature() - SimulationConstants.MIN_TEMPERATURE)
